Rename filenames to match classnames
