+++ assets/css/wcs-att-write-panels.css
