+++ assets/js/wcs-att-write-panels.js
