+++ includes/class-wcs-att-core-compatibility.php
