+++ includes/class-wcs-att-product.php
