+++ tests/framework/helpers/class-wcs-att-test-helpers-product.php
